add some files

+++ src/main/java/com/crypto/backendcrypto/repositories/AccountRepository.java
@@ -3,5 +3,5 @@ package com.crypto.backendcrypto.repositories;
 import com.crypto.backendcrypto.entitys.Account;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AccountRepository extends JpaRepository<Account, Integer> {
+public interface AccountRepository extends JpaRepository<Account, Long> {
 }
