+++ src/main/java/com/crypto/backendcrypto/repositories/CoinRepository.java
@@ -3,6 +3,6 @@ package com.crypto.backendcrypto.repositories;
 import com.crypto.backendcrypto.entitys.Coin;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoinRepository extends JpaRepository<Coin, Integer> {
+public interface CoinRepository extends JpaRepository<Coin, Long> {
 
 }
