+++ src/main/java/com/crypto/backendcrypto/repositories/TransactionRepository.java
@@ -3,6 +3,6 @@ package com.crypto.backendcrypto.repositories;
 import com.crypto.backendcrypto.entitys.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction,Integer> {
+public interface TransactionRepository extends JpaRepository<Transaction,Long> {
     
 }
